Add files via upload
The task is to define this nested list in a program and do the following data analysis.

a.) Compute the average number of sun hours for each month during the total data period (1929–2009).

b.) Which month has the best weather according to the means found in the preceding task?

c.) For each decade, 1930–1939, 1940–1949, : : :, 2000–2009, compute the aver- age number of sun hours per day in January and December. For example, use December 1949, January 1950, : : :, December 1958, and January 1959 as data for the decade 1950–1959. Are there any noticeable differences between the decades?

Link for the Data: https://github.com/hplgit/scipro-primer/blob/master/src/misc/Oxford_sun_hours.txt - (contains data of the num- ber of sun hours in Oxford, UK, for every month since January 1929.)

In reference to A Primer on Scientific Programming with Python Chapter 2.6.2: Analyzing List Data, the code here can by run on Python3